WEB/por_via_das_duvidas/challenge/requirements.txt
flask
flask_login
pycrypto
